Break up chapters duck
