+++ src/components/Reader/ReadingStatusUpdater.js
@@ -2,7 +2,7 @@
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
 import { withRouter } from "react-router-dom";
-import { updateReadingStatus } from "redux-ducks/chapters";
+import { updateReadingStatus } from "redux-ducks/chapters/actionCreators";
 
 // TODO: special case to consider?
 //       If you are on the last page of a chapter then go to the next chapter's first page,
